Refactor: Rename application from Sense/Veeda to Clair
This commit renames the application from its previous names (Sense and Veeda) to Clair across your entire codebase.

Changes include:

- **Asset Renaming:**
    - Logo files `sense_logo.png` renamed to `clair_logo.png` in mobile and web assets.

- **Mobile App (Flutter - `mobile_app`):**
    - `pubspec.yaml`: Updated app name to `clair_wallet`, description, author, and icon paths. Corrected all Dart import statements from `package:veeda_wallet/...` to `package:clair_wallet/...`.
    - `android/app/build.gradle.kts`: Updated `namespace` and `applicationId` to `com.clair.app`.
    - `android/app/src/main/AndroidManifest.xml`: Updated `android:label` to "Clair".
    - `ios/Runner/Info.plist`: Updated `CFBundleDisplayName` and `CFBundleName` to "Clair".
    - `ios/Runner.xcodeproj/project.pbxproj`: Updated `PRODUCT_BUNDLE_IDENTIFIER` from `dev.veeda.sense` to `app.clair.mobile`.
    - Updated app name in `lib/app.dart` and other UI-facing strings.
    - Updated Flutter web specific files in `mobile_app/web/` (manifest.json, index.html).

- **Backend (`backend`):**
    - `package.json`: Updated `name`, `description`, and `author`.
    - `wrangler.toml`: Updated application `name`.
    - `src/index.ts`: Updated API name in health check.
    - `src/middleware/cors.ts`: Updated allowed origins.
    - `create_backend_files.sh`: Updated generated CORS configuration.

- **Web App (Next.js - `web`):**
    - `package.json`: Updated `name`.
    - `app/layout.tsx`: Updated site metadata title to "Clair".
    - `messages/*.json`: Updated `app.title` in localization files.
    - `lib/apiClient.ts`: Updated fallback backend API URL.

- **Documentation:**
    - Updated app name in README files, design documents, implementation plans, and other `.md` files across the repository.

This comprehensive renaming ensures consistency for the "Clair" branding throughout your application stack and its documentation.
