+++ backend/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "veeda-wallet-backend",
+  "name": "clair-wallet-backend",
   "version": "1.0.0",
-  "description": "Backend API for Veeda Wallet spending tracker",
+  "description": "Backend API for Clair Wallet spending tracker",
   "main": "src/index.ts",
   "scripts": {
     "dev": "wrangler dev",
@@ -31,6 +31,6 @@
     "api",
     "supabase"
   ],
-  "author": "Veeda",
+  "author": "Clair",
   "license": "MIT"
 }
