+++ mobile_app/lib/core/di/isar_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:isar/isar.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:veeda_wallet/core/transaction_repository.dart'; // For TransactionSchema
+import 'package:clair_wallet/core/transaction_repository.dart'; // For TransactionSchema
 
 final isarProvider = FutureProvider<Isar>((ref) async {
   final dir = await getApplicationDocumentsDirectory();
