+++ mobile_app/lib/main.dart
@@ -3,5 +3,5 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'app.dart';
 
 void main() {
-  runApp(const ProviderScope(child: VeedaApp()));
+  runApp(const ProviderScope(child: ClairApp()));
 }
