+++ mobile_app/web/manifest.json
@@ -1,11 +1,11 @@
 {
-    "name": "mobile_app",
-    "short_name": "mobile_app",
+    "name": "Clair",
+    "short_name": "Clair",
     "start_url": ".",
     "display": "standalone",
     "background_color": "#0175C2",
     "theme_color": "#0175C2",
-    "description": "A new Flutter project.",
+    "description": "A personal finance app by Clair.",
     "orientation": "portrait-primary",
     "prefer_related_applications": false,
     "icons": [
