+++ web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "veeda-wallet",
+  "name": "clair-wallet",
   "version": "0.1.0",
   "private": true,
   "scripts": {
